#155 separate style for each component
