+++ frontend/components/form/InputEmail.css
@@ -1,0 +1,3 @@
+.main {
+    width: 100% !important;
+}
