+++ frontend/components/form/InputText.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {FormControl} from 'react-bootstrap';
-import styles from './index.css';
+import styles from './InputText.css';
 
 const InputText = ({input, placeholder}) => (
     <FormControl
         type="text"
-        className={styles.text}
+        className={styles.main}
         placeholder={placeholder}
         value={input.value}
         onChange={input.onChange}
