+++ frontend/components/form/InputTextArea.css
@@ -1,0 +1,4 @@
+.main {
+    width: 100% !important;
+    height: 10em !important;
+}
