+++ frontend/components/form/InputTextArea.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {FormControl} from 'react-bootstrap';
-import styles from './index.css';
+import styles from './InputTextArea.css';
 
 const InputTextArea = ({input}) => (
     <FormControl
         componentClass="textarea"
-        className={styles.textArea}
+        className={styles.main}
         value={input.value}
         onChange={input.onChange}
         onBlur={input.onBlur}
