+++ frontend/components/form/Required.css
@@ -1,0 +1,6 @@
+@import "../../variables";
+
+.main:after {
+    content: '*';
+    color: @brand-danger;
+}
