+++ frontend/components/form/SubmitButton.css
@@ -1,0 +1,8 @@
+.main {
+    display: flex !important;
+}
+
+.spinner {
+    margin-left: 5px;
+    color: white;
+}
