--- frontend/components/form/index.css
@@ -1,28 +1,0 @@
-@import "../../variables";
-
-.email {
-    width: 100% !important;
-}
-
-.textArea {
-    width: 100% !important;
-    height: 10em !important;
-}
-
-.text {
-    width: 100% !important;
-}
-
-.required:after {
-    content: '*';
-    color: @brand-danger;
-}
-
-.submitButton {
-    display: flex !important;
-}
-
-.submitSpinner {
-    margin-left: 5px;
-    color: white;
-}
