+++ frontend/components/index.less
@@ -1,7 +1,6 @@
 @import "../colors";
 @import "timeline/index";
 @import "statistics/index";
-@import "statement/index";
 @import (reference) "LeaderBoard";
 
 .leader-board.positive { .leader-board(@stats-positive); }
