+++ frontend/components/markdown/NegativeText.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import styles from './index.css';
+import styles from './NegativeText.css';
 
-const NegativeText = ({text}) => <span className={styles.negative}>{text}</span>;
+const NegativeText = ({text}) => <span className={styles.main}>{text}</span>;
 
 NegativeText.propTypes = {
     text: PropTypes.string.isRequired,
