+++ frontend/components/markdown/PositiveText.css
@@ -1,0 +1,5 @@
+@import "../../variables";
+
+.main {
+    color: @stats-positive;
+}
