+++ frontend/components/markdown/PositiveText.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import styles from './index.css';
+import styles from './PositiveText.css';
 
-const PositiveText = ({text}) => <span className={styles.positive}>{text}</span>;
+const PositiveText = ({text}) => <span className={styles.main}>{text}</span>;
 
 PositiveText.propTypes = {
     text: PropTypes.string.isRequired,
